refactor(ui): useKeyboardInsets is the suite's, not mail's
The calendar's event form is the second full-screen pane in the suite, and it
needs what mail's composer needed: the keyboard's insets, because iOS leaves the
layout viewport full-height and slides the visible part around underneath it.

The composable moves to src/composables unchanged. Its own explanation of why
each inset is load-bearing goes with it — that is the part a second reader of
this problem needs, and it was sitting in mail's utils.
