+++ frontend/src/apps/mail/pages/ComposeView.vue
@@ -265,7 +265,8 @@ import { Button, Combobox } from 'frappe-ui'
 import { Icon as FeatherIcon, TextEditor, TextEditorFixedMenu } from 'frappe-ui/experimental'
 
 import { formatBytes, raiseToast } from '@/apps/mail/utils'
-import { useKeyboardInsets, useTextEditorButtons } from '@/apps/mail/utils/composables'
+import { useKeyboardInsets } from '@/composables/useKeyboardInsets'
+import { useTextEditorButtons } from '@/apps/mail/utils/composables'
 import { CustomImageExtension, CustomParagraphExtension, uploadFunction } from '@/apps/mail/utils/text-editor'
 import { takePendingCompose } from '@/apps/mail/composables/composeHandoff'
 import { useComposeMail } from '@/apps/mail/composables/useComposeMail'
